refactor(qna): 폴더 구조 올바르게 변경
